Fix warnings "implicit conversion shortens 64-bit value into a 32-bit value"

+++ libyara/hex_grammar.h
@@ -56,7 +56,7 @@
 typedef union YYSTYPE
 #line 64 "hex_grammar.y"
 {
-  int integer;
+  int64_t integer;
   RE_NODE *re_node;
 }
 /* Line 1529 of yacc.c.  */
